+++ ferry_test_graphql/lib/queries/interfaces/hero_for_episode.req.gql.dart
@@ -2,7 +2,7 @@
 
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
-import 'package:ferry/ferry.dart' as _i1;
+import 'package:ferry_exec/ferry_exec.dart' as _i1;
 import 'package:ferry_test_graphql/queries/interfaces/hero_for_episode.ast.gql.dart'
     as _i5;
 import 'package:ferry_test_graphql/queries/interfaces/hero_for_episode.data.gql.dart'
@@ -28,7 +28,8 @@ abstract class GHeroForEpisodeReq
   static void _initializeBuilder(GHeroForEpisodeReqBuilder b) => b
     ..operation =
         _i4.Operation(document: _i5.document, operationName: 'HeroForEpisode')
-    ..requestId = _i6.Uuid().v1();
+    ..requestId = _i6.Uuid().v1()
+    ..executeOnListen = true;
   _i3.GHeroForEpisodeVars get vars;
   _i4.Operation get operation;
   _i4.Request get execRequest =>
@@ -47,6 +48,8 @@ abstract class GHeroForEpisodeReq
   Map<String, dynamic> get updateCacheHandlerContext;
   @nullable
   _i1.FetchPolicy get fetchPolicy;
+  @nullable
+  bool get executeOnListen;
   @override
   _i2.GHeroForEpisodeData parseData(Map<String, dynamic> json) =>
       _i2.GHeroForEpisodeData.fromJson(json);