+++ normalize/lib/src/denormalize_fragment.dart
@@ -4,6 +4,7 @@ import 'package:normalize/normalize.dart';
 
 import 'package:normalize/src/config/normalization_config.dart';
 import 'package:normalize/src/policies/type_policy.dart';
+import 'package:normalize/src/utils/get_fragment_map.dart';
 import 'package:normalize/src/utils/resolve_data_id.dart';
 import 'package:normalize/src/utils/add_typename_visitor.dart';
 import 'package:normalize/src/utils/exceptions.dart';
@@ -34,6 +35,7 @@ Map<String, dynamic> denormalizeFragment({
   DataIdResolver dataIdFromObject,
   bool addTypename = false,
   bool returnPartialData = false,
+  bool handleException = true,
   String referenceKey = '\$ref',
 }) {
   if (addTypename) {
@@ -43,11 +45,7 @@ Map<String, dynamic> denormalizeFragment({
     );
   }
 
-  final fragmentMap = {
-    for (var fragmentDefinition
-        in document.definitions.whereType<FragmentDefinitionNode>())
-      fragmentDefinition.name.value: fragmentDefinition
-  };
+  final fragmentMap = getFragmentMap(document);
 
   if (fragmentMap.length > 1 && fragmentName == null) {
     throw Exception(
@@ -82,6 +80,7 @@ Map<String, dynamic> denormalizeFragment({
     fragmentMap: fragmentMap,
     dataIdFromObject: dataIdFromObject,
     addTypename: addTypename,
+    allowPartialData: returnPartialData,
   );
 
   try {
@@ -89,9 +88,11 @@ Map<String, dynamic> denormalizeFragment({
       selectionSet: fragmentDefinition.selectionSet,
       dataForNode: read(dataId),
       config: config,
-      returnPartialData: returnPartialData,
     );
   } on PartialDataException {
-    return null;
+    if (handleException) {
+      return null;
+    }
+    rethrow;
   }
 }