+++ ferry_hive_store/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2020-10-16
+
+### Changed
+
+- upgrade `rxdart` to latest version
+
+## [0.3.0] - 2020-10-07
+
+### Changed
+
+- watch now only watches a specific `dataId`
+
 ## [0.2.2] - 2020-09-09
 
 ### Added