+++ normalize/CHANGELOG.md
@@ -7,6 +7,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.3-alpha] - 2020-11-06
+
+### Added
+
+- add a `path` to `PartialDataException` to surface invalid paths
+- add `acceptPartialData` option to `normalizeFragment` and `normalizeOperation` for rejecting partial `data`
+- add `handleException` option to `denormalizeFragment` and `denormalizeOperation` for throwing a `PartialDataException` instead of returning `null` on partial `data`
+- add a `validateOperationDataStructure` utility for validating structure against a document
+
+## [0.4.2] - 2020-10-07
+
+### Changed
+
+- rename `reachableDataIds` to `reachableIds`
+
+### Added
+
+- add `operationFieldNames` method
+
+## [0.4.1] - 2020-09-22
+
+### Changed
+
+- prevent subscription from adding `__typename`
+
 ## [0.4.0] - 2020-09-10
 
 ### Changed
@@ -87,7 +112,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - add to ferry monorepo
-
 
 ## [0.1.4] - 2020-02-22